Organização de pastas e adição de variantes.
