+++ Services/Infrastructure/Supabase/Models/SupabaseVariant.cs
@@ -1,14 +1,9 @@
 ﻿using OrchidPro.Models;
-using OrchidPro.Services.Base;
 using Supabase.Postgrest.Attributes;
 using Supabase.Postgrest.Models;
 
-namespace OrchidPro.Services.Data;
+namespace OrchidPro.Services.Infrastructure.Supabase.Models;
 
-/// <summary>
-/// Supabase database model representing variants table.
-/// Maps between database schema and application domain models.
-/// </summary>
 [Table("variants")]
 public class SupabaseVariant : BaseModel
 {
@@ -39,9 +34,6 @@ public class SupabaseVariant : BaseModel
     [Column("sync_hash")]
     public string? SyncHash { get; set; }
 
-    /// <summary>
-    /// Convert SupabaseVariant to domain Variant model
-    /// </summary>
     public Variant ToVariant()
     {
         return new Variant
@@ -58,9 +50,6 @@ public class SupabaseVariant : BaseModel
         };
     }
 
-    /// <summary>
-    /// Convert domain Variant model to SupabaseVariant
-    /// </summary>
     public static SupabaseVariant FromVariant(Variant variant)
     {
         return new SupabaseVariant
@@ -76,21 +65,4 @@ public class SupabaseVariant : BaseModel
             SyncHash = variant.SyncHash
         };
     }
-}
-
-/// <summary>
-/// REFACTORED: Variant service using BaseSupabaseEntityService.
-/// Reduced from ~150 lines to minimal implementation focused on Variant-specific logic.
-/// MODERNIZED: Uses primary constructor for cleaner initialization.
-/// </summary>
-public class SupabaseVariantService(SupabaseService supabaseService) : BaseSupabaseEntityService<Variant, SupabaseVariant>(supabaseService)
-{
-    protected override string EntityTypeName => "Variant";
-    protected override string EntityPluralName => "Variants";
-
-    protected override Variant ConvertToEntity(SupabaseVariant supabaseModel)
-        => supabaseModel.ToVariant();
-
-    protected override SupabaseVariant ConvertFromEntity(Variant entity)
-        => SupabaseVariant.FromVariant(entity);
 }
