+++ Services/Infrastructure/Supabase/Repositories/SupabaseGenusRepository.cs
@@ -1,38 +1,26 @@
 ﻿using OrchidPro.Models;
 using OrchidPro.Models.Base;
 using OrchidPro.Services.Base;
+using OrchidPro.Services.Contracts;
+using OrchidPro.Services.Infrastructure.Supabase.Models;
 using OrchidPro.Services.Data;
 using OrchidPro.Extensions;
 
-namespace OrchidPro.Services;
+namespace OrchidPro.Services.Infrastructure.Supabase.Repositories;
 
-/// <summary>
-/// ULTRA MINIMAL Genus repository - just the absolute essentials!
-/// BaseHierarchicalRepository handles ALL operations - this only provides service connection.
-/// CRITICAL: Always loads with Family data for immediate display (no "Unknown Family" delays).
-/// </summary>
-public class GenusRepository : BaseHierarchicalRepository<Genus, Family>, IGenusRepository
+public class SupabaseGenusRepository(SupabaseService supabaseService, IFamilyRepository familyRepository)
+    : BaseHierarchicalRepository<Genus, Family>(supabaseService, familyRepository), IGenusRepository
 {
-    #region Private Fields
-
-    private readonly SupabaseGenusService _genusService;
-
-    #endregion
-
-    #region Required Base Implementation
+    private readonly BaseSupabaseEntityService<Genus, SupabaseGenus> _supabaseEntityService = new InternalSupabaseGenusService(supabaseService);
 
     protected override string EntityTypeName => "Genus";
     protected override string ParentEntityTypeName => "Family";
 
-    /// <summary>
-    /// CRITICAL: Always load with Family data immediately (like Species with Genus)
-    /// Eliminates "Unknown Family" delays by populating parent data upfront
-    /// </summary>
     protected override async Task<IEnumerable<Genus>> GetAllFromServiceAsync()
     {
-        var rawGenera = await _genusService.GetAllAsync();
+        var rawGenera = await _supabaseEntityService.GetAllAsync();
         var generaWithFamily = await PopulateParentDataAsync([.. rawGenera]);
-        this.LogInfo($"✅ Loaded {generaWithFamily.Count} genera WITH family data immediately");
+        this.LogInfo($"Loaded {generaWithFamily.Count} genera WITH family data immediately");
         return generaWithFamily;
     }
 
@@ -43,13 +31,13 @@ public class GenusRepository : BaseHierarchicalRepository<Genus, Family>, IGenus
     }
 
     protected override async Task<Genus?> CreateInServiceAsync(Genus entity)
-        => await _genusService.CreateAsync(entity);
+        => await _supabaseEntityService.CreateAsync(entity);
 
     protected override async Task<Genus?> UpdateInServiceAsync(Genus entity)
-        => await _genusService.UpdateAsync(entity);
+        => await _supabaseEntityService.UpdateAsync(entity);
 
     protected override async Task<bool> DeleteInServiceAsync(Guid id)
-        => await _genusService.DeleteAsync(id);
+        => await _supabaseEntityService.DeleteAsync(id);
 
     protected override async Task<bool> NameExistsInServiceAsync(string name, Guid? excludeId)
     {
@@ -58,93 +46,40 @@ public class GenusRepository : BaseHierarchicalRepository<Genus, Family>, IGenus
             string.Equals(g.Name, name, StringComparison.OrdinalIgnoreCase) && g.Id != excludeId);
     }
 
-    #endregion
-
-    #region Constructor
-
-    public GenusRepository(
-        SupabaseService supabaseService,
-        SupabaseGenusService genusService,
-        IFamilyRepository familyRepository)
-        : base(supabaseService, familyRepository)
-    {
-        _genusService = genusService ?? throw new ArgumentNullException(nameof(genusService));
-        this.LogInfo("ULTRA MINIMAL GenusRepository - immediate family loading enabled!");
-    }
-
-    #endregion
-
-    #region IGenusRepository - Required Interface Methods
-
-    /// <summary>
-    /// Get genera by family - uses base GetByParentIdAsync
-    /// </summary>
+    // IGenusRepository implementations
     public async Task<List<Genus>> GetByFamilyIdAsync(Guid familyId, bool includeInactive = false)
         => await GetByParentIdAsync(familyId, includeInactive);
 
-    /// <summary>
-    /// Get filtered genera by family - uses base GetFilteredByParentAsync
-    /// </summary>
     public async Task<List<Genus>> GetFilteredByFamilyAsync(Guid familyId, string? searchText = null, bool? statusFilter = null)
         => await GetFilteredByParentAsync(familyId, searchText, statusFilter);
 
-    /// <summary>
-    /// Get count by family - uses base GetCountByParentAsync
-    /// </summary>
     public async Task<int> GetCountByFamilyAsync(Guid familyId, bool includeInactive = false)
         => await GetCountByParentAsync(familyId, includeInactive);
 
-    /// <summary>
-    /// Check name exists in family - uses base NameExistsInParentAsync
-    /// </summary>
     public async Task<bool> NameExistsInFamilyAsync(string name, Guid familyId, Guid? excludeId = null)
         => await NameExistsInParentAsync(name, familyId, excludeId);
 
-    /// <summary>
-    /// Validate family access - uses base ValidateParentAccessAsync
-    /// </summary>
     public async Task<bool> ValidateFamilyAccessAsync(Guid familyId)
         => await ValidateParentAccessAsync(familyId);
 
-    /// <summary>
-    /// Populate family data - uses base PopulateParentDataAsync
-    /// </summary>
     public async Task<List<Genus>> PopulateFamilyDataAsync(List<Genus> genera)
         => await PopulateParentDataAsync(genera);
 
-    /// <summary>
-    /// Get all genera with family data - uses base GetAllWithParentAsync
-    /// </summary>
     public async Task<List<Genus>> GetAllWithFamilyAsync(bool includeInactive = false)
         => await GetAllWithParentAsync(includeInactive);
 
-    /// <summary>
-    /// Get filtered with family - uses base GetFilteredWithParentAsync
-    /// </summary>
     public async Task<List<Genus>> GetFilteredWithFamilyAsync(string? searchText = null, bool? statusFilter = null, Guid? familyId = null)
         => await GetFilteredWithParentAsync(searchText, statusFilter, familyId);
 
-    /// <summary>
-    /// Delete by family - uses base DeleteByParentAsync
-    /// </summary>
     public async Task<int> DeleteByFamilyAsync(Guid familyId)
         => await DeleteByParentAsync(familyId);
 
-    /// <summary>
-    /// Bulk update family - uses base BulkUpdateParentAsync
-    /// </summary>
     public async Task<int> BulkUpdateFamilyAsync(List<Guid> genusIds, Guid newFamilyId)
         => await BulkUpdateParentAsync(genusIds, newFamilyId);
 
-    /// <summary>
-    /// Get statistics by family - uses base GetStatisticsByParentAsync
-    /// </summary>
     public async Task<BaseStatistics> GetStatisticsByFamilyAsync(Guid familyId)
         => await GetStatisticsByParentAsync(familyId);
 
-    /// <summary>
-    /// Get genus statistics
-    /// </summary>
     public async Task<GenusStatistics> GetGenusStatisticsAsync()
     {
         var hierarchicalStats = await GetHierarchicalStatisticsAsync();
@@ -164,6 +99,17 @@ public class GenusRepository : BaseHierarchicalRepository<Genus, Family>, IGenus
             FamilyDistribution = hierarchicalStats.ParentDistribution
         };
     }
+}
 
-    #endregion
+internal class InternalSupabaseGenusService(SupabaseService supabaseService)
+    : BaseSupabaseEntityService<Genus, SupabaseGenus>(supabaseService)
+{
+    protected override string EntityTypeName => "Genus";
+    protected override string EntityPluralName => "Genera";
+
+    protected override Genus ConvertToEntity(SupabaseGenus supabaseModel)
+        => supabaseModel.ToGenus();
+
+    protected override SupabaseGenus ConvertFromEntity(Genus entity)
+        => SupabaseGenus.FromGenus(entity);
 }
