--- Services/SupabaseSpeciesService.cs
@@ -1,324 +1,0 @@
-﻿// ==========================================
-using OrchidPro.Models;
-using OrchidPro.Models.Base;
-using OrchidPro.Services.Data;
-using OrchidPro.Services.Base;
-using Supabase.Postgrest.Attributes;
-using Supabase.Postgrest.Models;
-using OrchidPro.Extensions;
-
-namespace OrchidPro.Services;
-
-/// <summary>
-/// Supabase database model representing species table with genus relationship.
-/// Maps between database schema and application domain models.
-/// </summary>
-[Table("species")]
-public class SupabaseSpecies : BaseModel
-{
-    [PrimaryKey("id", false)]
-    public Guid Id { get; set; } = Guid.NewGuid();
-
-    [Column("genus_id")]
-    public Guid GenusId { get; set; }
-
-    [Column("user_id")]
-    public Guid? UserId { get; set; }
-
-    [Column("name")]
-    public string Name { get; set; } = string.Empty;
-
-    [Column("description")]
-    public string? Description { get; set; }
-
-    [Column("is_active")]
-    public bool? IsActive { get; set; } = true;
-
-    [Column("is_favorite")]
-    public bool? IsFavorite { get; set; } = false;
-
-    [Column("created_at")]
-    public DateTime? CreatedAt { get; set; } = DateTime.UtcNow;
-
-    [Column("updated_at")]
-    public DateTime? UpdatedAt { get; set; } = DateTime.UtcNow;
-
-    [Column("common_name")]
-    public string? CommonName { get; set; }
-
-    [Column("cultivation_notes")]
-    public string? CultivationNotes { get; set; }
-
-    [Column("habitat_info")]
-    public string? HabitatInfo { get; set; }
-
-    [Column("flowering_season")]
-    public string? FloweringSeason { get; set; }
-
-    [Column("flower_colors")]
-    public string? FlowerColors { get; set; }
-
-    [Column("size_category")]
-    public string? SizeCategory { get; set; } = "Medium";
-
-    [Column("rarity_status")]
-    public string? RarityStatus { get; set; } = "Common";
-
-    [Column("fragrance")]
-    public bool? Fragrance { get; set; } = false;
-
-    [Column("temperature_preference")]
-    public string? TemperaturePreference { get; set; }
-
-    [Column("light_requirements")]
-    public string? LightRequirements { get; set; }
-
-    [Column("humidity_preference")]
-    public string? HumidityPreference { get; set; }
-
-    [Column("growth_habit")]
-    public string? GrowthHabit { get; set; }
-
-    [Column("bloom_duration")]
-    public string? BloomDuration { get; set; }
-
-    /// <summary>
-    /// Convert SupabaseSpecies to domain Species model
-    /// </summary>
-    public Species ToSpecies()
-    {
-        return new Species
-        {
-            Id = this.Id,
-            GenusId = this.GenusId,
-            UserId = this.UserId,
-            Name = this.Name ?? string.Empty,
-            Description = this.Description,
-            IsActive = this.IsActive ?? true,
-            IsFavorite = this.IsFavorite ?? false,
-            CreatedAt = this.CreatedAt ?? DateTime.UtcNow,
-            UpdatedAt = this.UpdatedAt ?? DateTime.UtcNow,
-            CommonName = this.CommonName,
-            CultivationNotes = this.CultivationNotes,
-            HabitatInfo = this.HabitatInfo,
-            FloweringSeason = this.FloweringSeason,
-            FlowerColors = this.FlowerColors,
-            SizeCategory = this.SizeCategory ?? "Medium",
-            RarityStatus = this.RarityStatus ?? "Common",
-            Fragrance = this.Fragrance ?? false,
-            TemperaturePreference = this.TemperaturePreference,
-            LightRequirements = this.LightRequirements,
-            HumidityPreference = this.HumidityPreference,
-            GrowthHabit = this.GrowthHabit,
-            BloomDuration = this.BloomDuration
-        };
-    }
-
-    /// <summary>
-    /// Convert domain Species model to SupabaseSpecies
-    /// </summary>
-    public static SupabaseSpecies FromSpecies(Species species)
-    {
-        return new SupabaseSpecies
-        {
-            Id = species.Id,
-            GenusId = species.GenusId,
-            UserId = species.UserId,
-            Name = species.Name,
-            Description = species.Description,
-            IsActive = species.IsActive,
-            IsFavorite = species.IsFavorite,
-            CreatedAt = species.CreatedAt,
-            UpdatedAt = species.UpdatedAt,
-            CommonName = species.CommonName,
-            CultivationNotes = species.CultivationNotes,
-            HabitatInfo = species.HabitatInfo,
-            FloweringSeason = species.FloweringSeason,
-            FlowerColors = species.FlowerColors,
-            SizeCategory = species.SizeCategory,
-            RarityStatus = species.RarityStatus,
-            Fragrance = species.Fragrance,
-            TemperaturePreference = species.TemperaturePreference,
-            LightRequirements = species.LightRequirements,
-            HumidityPreference = species.HumidityPreference,
-            GrowthHabit = species.GrowthHabit,
-            BloomDuration = species.BloomDuration
-        };
-    }
-}
-
-/// <summary>
-/// REFACTORED: Species service using BaseSupabaseEntityService.
-/// Reduced from ~200 lines to minimal implementation focused on Species-specific logic.
-/// 🚀 NEW: Added optimized JOIN query method to eliminate N+1 query problem.
-/// </summary>
-public class SupabaseSpeciesService(SupabaseService supabaseService) : BaseSupabaseEntityService<Species, SupabaseSpecies>(supabaseService)
-{
-    protected override string EntityTypeName => "Species";
-    protected override string EntityPluralName => "Species";
-
-    protected override Species ConvertToEntity(SupabaseSpecies supabaseModel)
-        => supabaseModel.ToSpecies();
-
-    protected override SupabaseSpecies ConvertFromEntity(Species entity)
-        => SupabaseSpecies.FromSpecies(entity);
-
-    /// <summary>
-    /// 🚀 NEW: Optimized method using JOIN query instead of N+1 queries
-    /// Expected performance: 300ms instead of 2500ms for loading species with related data
-    /// </summary>
-    /// <summary>
-    /// 🚀 NEW: Optimized method using JOIN query instead of N+1 queries
-    /// Expected performance: 300ms instead of 2500ms for loading species with related data
-    /// </summary>
-    public async Task<IEnumerable<Species>> GetAllWithJoinAsync()
-    {
-        var result = await this.SafeDataExecuteAsync(async () =>
-        {
-            if (_supabaseService.Client == null)
-                return [];
-
-            this.LogInfo("🚀 OPTIMIZED: Loading species with PostgreSQL JOIN query");
-
-            var currentUserId = GetCurrentUserId();
-
-            try
-            {
-                // PostgREST JOIN syntax - single query for all related data
-                var response = await _supabaseService.Client
-                    .From<SupabaseSpecies>()
-                    .Select(@"
-                    *,
-                    genus:genus_id(
-                        id, name, family_id, description, is_active, is_favorite, created_at, updated_at, user_id,
-                        family:family_id(id, name, description, is_active, is_favorite, created_at, updated_at, user_id)
-                    )
-                ")
-                    .Get();
-
-                if (response?.Models == null)
-                {
-                    this.LogInfo("No species models returned from JOIN query");
-                    return [];
-                }
-
-                // FIXED CA1829: Use Count property instead of Count() method
-                this.LogInfo($"🚀 OPTIMIZED: Retrieved {response.Models.Count} species records with JOIN data");
-
-                // Filter: user entities OR system defaults (UserId == null)
-                var filteredModels = response.Models.Where(model =>
-                    GetModelUserId(model) == currentUserId || GetModelUserId(model) == null);
-
-                // Convert and order species - using traditional syntax for compatibility
-                var species = filteredModels
-                    .Select(ConvertToEntity)
-                    .OrderBy(entity => GetEntityName(entity))
-                    .ToList();
-
-                this.LogInfo($"🚀 OPTIMIZED: Successfully converted {species.Count} species with complete hierarchy");
-
-                return species;
-            }
-            catch (Exception ex)
-            {
-                this.LogError(ex, "Error in JOIN query, falling back to basic query");
-
-                // Fallback to basic query if JOIN fails
-                return await GetAllAsync();
-            }
-
-        }, EntityPluralName);
-
-        return result.Success && result.Data != null ? result.Data : [];
-    }
-
-    /// <summary>
-    /// 🚀 SIMPLER: Batch loading approach - more reliable than JOIN
-    /// Uses 2-3 simple queries instead of complex JOIN syntax
-    /// </summary>
-    public async Task<IEnumerable<Species>> GetAllWithBatchLoadingAsync()
-    {
-        var result = await this.SafeDataExecuteAsync(async () =>
-        {
-            if (_supabaseService.Client == null)
-                return [];
-
-            this.LogInfo("🚀 BATCH: Loading species with 3-query batch approach");
-
-            var currentUserId = GetCurrentUserId();
-
-            // Query 1: Get all Species
-            var speciesResponse = await _supabaseService.Client
-                .From<SupabaseSpecies>()
-                .Select("*")
-                .Get();
-
-            if (speciesResponse?.Models == null)
-                return Enumerable.Empty<Species>();
-
-            var filteredSpecies = speciesResponse.Models.Where(model =>
-                GetModelUserId(model) == currentUserId || GetModelUserId(model) == null);
-
-            // Convert species with explicit type declaration for collection expression
-            var species = filteredSpecies.Select(ConvertToEntity).ToList();
-
-            // FIXED CA1860: Use Count property instead of Any() for better performance
-            if (species.Count == 0)
-                return species;
-
-            this.LogInfo($"🚀 BATCH: Got {species.Count} species, now getting genus data");
-
-            // Query 2: Get unique Genus IDs in batch  
-            // FIXED IDE0301: Use ToArray() for Supabase filter compatibility
-            var genusIds = species.Select(s => s.GenusId).Distinct().ToArray();
-
-            var genusResponse = await _supabaseService.Client
-                .From<SupabaseGenus>()
-                .Select("*")
-                .Filter("id", Supabase.Postgrest.Constants.Operator.In, genusIds)
-                .Get();
-
-            if (genusResponse?.Models != null)
-            {
-                // FIXED CA1829: Use Count property instead of Count() method
-                this.LogInfo($"🚀 BATCH: Got {genusResponse.Models.Count} genus records");
-
-                // FIXED CS8073: Remove unnecessary null check for Guid value type
-                // FIXED IDE0301: Simplified dictionary initialization using collection expressions
-                var genusLookup = genusResponse.Models.ToDictionary(
-                    g => Guid.Parse(g.Id.ToString()!),
-                    g => new Genus
-                    {
-                        Id = Guid.Parse(g.Id.ToString()!),
-                        FamilyId = Guid.Parse(g.FamilyId.ToString()!),
-                        Name = g.Name?.ToString() ?? "",
-                        Description = g.Description?.ToString(),
-                        IsActive = g.IsActive ?? true,
-                        IsFavorite = g.IsFavorite ?? false,
-                        CreatedAt = g.CreatedAt ?? DateTime.UtcNow,
-                        UpdatedAt = g.UpdatedAt ?? DateTime.UtcNow,
-                        UserId = g.UserId != null ? Guid.Parse(g.UserId.ToString()!) : null
-                    }
-                );
-
-                // Populate genus data
-                foreach (var spec in species)
-                {
-                    if (genusLookup.TryGetValue(spec.GenusId, out var genus))
-                    {
-                        spec.Genus = genus;
-                    }
-                }
-
-                var populatedCount = species.Count(s => s.Genus != null);
-                this.LogInfo($"🚀 BATCH: Populated {populatedCount}/{species.Count} species with genus data");
-            }
-
-            // Return ordered species as IEnumerable
-            return species.OrderBy(s => s.Name);
-
-        }, EntityPluralName);
-
-        return result.Success && result.Data != null ? result.Data : [];
-    }
-}
