+++ Views/Pages/Botanical/FamiliesListPage.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using OrchidPro.Extensions;
-using OrchidPro.ViewModels.Families;
+using OrchidPro.ViewModels.Botanical.Families;
 using OrchidPro.Views.Base;
 
 namespace OrchidPro.Views.Pages.Botanical;
