+++ Views/Pages/Botanical/SpeciesEditPage.xaml
@@ -2,7 +2,7 @@
 <ContentPage x:Class="OrchidPro.Views.Pages.Botanical.SpeciesEditPage"
              xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:vm="clr-namespace:OrchidPro.ViewModels.Species"
+             xmlns:vm="clr-namespace:OrchidPro.ViewModels.Botanical.Species"
              xmlns:sfbusy="clr-namespace:Syncfusion.Maui.Core;assembly=Syncfusion.Maui.Core"
              xmlns:sfcombobox="clr-namespace:Syncfusion.Maui.Inputs;assembly=Syncfusion.Maui.Inputs"
              x:DataType="vm:SpeciesEditViewModel"
