+++ Views/Pages/Botanical/SpeciesListPage.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using OrchidPro.Extensions;
-using OrchidPro.ViewModels.Species;
+using OrchidPro.ViewModels.Botanical.Species;
 using OrchidPro.Views.Base;
 
 namespace OrchidPro.Views.Pages.Botanical;
